Remove 'profile_limiters' and fuse 'set_inner_as_kord' stencils
